requirements.txt
streamlit
python-docx
python-dotenv
google-generativeai
PyPDF2
